requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.16
django-appconf==1.0.4
django-cascading-dropdown-widget==0.2.1
django-compressor==2.4
django-crispy-forms==1.9.2
django-filter==2.3.0
django-navutils==0.7
django-nvd3==0.9.7
django-phonenumber-field==5.0.0
django-postgres-extensions==0.9.3
django-redis==4.12.1
django-rest-swagger==2.2.0
django-select2==7.4.2
django-static-arrive==2.4.1.0
django-static-jquery3==5.0.0
djangorestframework==3.11.1
djangorestframework-gis==0.16
djangorestframework-jsonapi==3.2.0
entrypoints==0.3
geographiclib==1.50
geopy==2.0.0
idna==2.10
inflection==0.5.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
openapi-codec==1.3.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
persisting-theory==0.2.1
pexpect==4.8.0
phonenumbers==8.12.9
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
rcssmin==1.0.6
redis==3.5.3
requests==2.24.0
rjsmin==1.1.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
